refactor(estimation): fold the result-accessor mixin into BaseRegression
`ResultAccessorMixin` existed only to lend `Feols` its result accessors, and it
carried stub declarations for attributes and guards the host class had to
supply. Folding it into `BaseRegression` removes the indirection: `tidy()`,
`confint()`, `resid()`, `get_inference()`, `get_performance()`, the report
bindings, and the SAVI accessors now live on the class that owns the state they
read, and the declared attributes become annotated fields with real types
instead of host-class promises.

`TidyColumnAccessors` stays a mixin, in `models/_tidy_accessors.py`, because
`FixestMulti` shares it.
